requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
blessings==1.7
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
clarifai==2.6.1
configparser==3.7.4
cryptography==2.6.1
decorator==4.4.0
EasyProcess==0.2.5
emoji==0.5.2
fake-useragent==0.1.11
future==0.17.1
fuzzywuzzy==0.17.0
googleapis-common-protos==1.5.10
grpcio==1.20.1
haversine==2.1.1
idna==2.8
instabot-py==0.4.7
instaloader==4.2.1
instapy==0.4.3
instapy-chromedriver==2.46
jsonschema==3.0.1
MeaningCloud-python==1.1.1
networkx==2.3
numpy==1.16.4
pandas==0.24.2
Pillow==6.0.0
plyer==1.4.0
protobuf==3.7.1
pycparser==2.19
pyrsistent==0.15.2
python-dateutil==2.8.0
python-Levenshtein==0.12.0
python-telegram-bot==11.1.0
pytz==2019.1
PyVirtualDisplay==0.2.1
PyYAML==5.1
regex==2019.4.14
requests==2.21.0
selenium==3.141.0
six==1.12.0
sleekxmpp==1.3.3
staticmap==0.5.4
units==0.7
urllib3==1.24.3
